+++ workers/unity/Packages/io.improbable.gdk.buildsystem/.codegen/Source/WorkerGenerationJob.cs
@@ -18,8 +18,8 @@ namespace Improbable.Gdk.CodeGenerator
         private readonly string relativeOutputPath = Path.Combine("improbable", "buildsystem");
         private readonly string relativeEditorPath = Path.Combine("improbable", "buildsystem", "Editor");
 
-        public WorkerGenerationJob(string baseOutputDir, IFileSystem fileSystem, DetailsStore detailsStore, bool force)
-            : base(baseOutputDir, fileSystem, detailsStore, force)
+        public WorkerGenerationJob(CodegenJobOptions options, IFileSystem fileSystem, DetailsStore detailsStore)
+            : base(options.AsEditor(), fileSystem, detailsStore)
         {
             const string jobName = nameof(WorkerGenerationJob);
             Logger.Trace($"Initialising {jobName}.");